Inventory for player
This is a file that lets a player make an inventory list and save it to a text file. Once saved it will be able to be opened. Only one can be saved before being overwritten. Future plans are to make it so the text file can be read in as a list again to be edited